requirements.txt
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Django==4.2.2
django-extensions==3.2.3
django-filter==24.2
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
idna==3.7
inflection==0.5.1
ipython==8.24.0
jedi==0.19.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.13.38
phonenumberslite==8.13.38
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
stripe==9.12.0
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
